html css
Added in homepage grid +

Removed green links throughout
Change Pinterest back to grayscale brand colours
Redefined all links structure css into grey
Swapped out logo change on scroll for the D+G Slimline version
Removed grey background throughout, and set to menu white
I have extended the width of the website container. We CAN NOT place the
logos and menus HARD to the left or HARD to the right as you have
requested and is present on your old site. This is bad practice, and is
in contradiction to both the example websites you have sent, which do
not fill to the hard side of the page. They have a container, this is
vital. However I have made our container much wider than previously and
this looks great and will achieve your wishes of having a wider site.
However as we write the code, we must limit the width or set a maximum,
otherwise we lose control and we can not possible restructure it
depending on screen size etc.
Request a cat font was incorrect, well spotted. This has been updated.
Tidied up header with address and request links similar to Michaelis
Boyd
Added in additional menu items
Remove the dots
Reduced the height of the dropdowns by 50%
Create dropdown display on hover event rather than click event
Re-ordered the images as per the project page. I really think that you
need the spaces between them, Michaelis Boyd does it, and I did it and
it looked unprofessional. The images contrast and do not sit nicely, the
gaps make it look more editorial and neater. (happy to change if you
insist of course). I also left two large ones in the middle as I think
it breaks up the monotony of just regulated squares. But again happy to
change to above as requested. I didn’t change the images around as I am
out of high res ones, so I need ALL your images in the highest
resolution you have and instructions as to which one belongs to each
project.